[ADD] Responsivity

+++ Hider/Hider/TransformUI.cpp
@@ -4,17 +4,17 @@
 TransformUI::TransformUI()
     : positionX(0), positionY(0), width(100), height(100), anchorX(0.5f), anchorY(0.5f) {}
 
-TransformUI::TransformUI(int posX, int posY, int w, int h, float ancX, float ancY)
+TransformUI::TransformUI(float posX, float posY, float w, float h, float ancX, float ancY)
     : positionX(posX), positionY(posY), width(w), height(h), anchorX(ancX), anchorY(ancY) {}
 
-int TransformUI::getPositionX() const { return positionX; }
-int TransformUI::getPositionY() const { return positionY; }
+float TransformUI::getPositionX() const { return positionX; }
+float TransformUI::getPositionY() const { return positionY; }
 float TransformUI::getAnchorX() const { return anchorX; }
 float TransformUI::getAnchorY() const { return anchorY; }
-int TransformUI::getWidth() const { return width; }
-int TransformUI::getHeight() const { return height; }
+float TransformUI::getWidth() const { return width; }
+float TransformUI::getHeight() const { return height; }
 
-void TransformUI::setPosition(int x, int y) {
+void TransformUI::setPosition(float x, float y) {
     positionX = x;
     positionY = y;
 }
@@ -24,15 +24,15 @@ void TransformUI::setAnchor(float x, float y) {
     anchorY = y;
 }
 
-void TransformUI::setSize(int w, int h) {
+void TransformUI::setSize(float w, float h) {
     width = w;
     height = h;
 }
 
-int TransformUI::newGetPositionX() const {
+float TransformUI::newGetPositionX() const {
     return positionX - (anchorX * width);
 }
 
-int TransformUI::newGetPositionY() const {
+float TransformUI::newGetPositionY() const {
     return positionY - (anchorY * height);
 }
